+++ src/main/java/mediathek/update/UpdateNotificationDialog.kt
@@ -1,70 +1,84 @@
-package mediathek.update;
+package mediathek.update
 
-import javafx.application.Platform;
-import javafx.scene.Scene;
-import javafx.scene.web.WebEngine;
-import javafx.scene.web.WebView;
-import mediathek.config.Konstanten;
-import mediathek.gui.actions.DisposeDialogAction;
-import mediathek.gui.dialog.ButtonFlowPanel;
-import mediathek.gui.dialog.ButtonPanel;
-import mediathek.tool.EscapeKeyHandler;
-import mediathek.tool.GuiFunktionen;
-import mediathek.tool.Version;
+import javafx.application.Platform
+import javafx.scene.Scene
+import javafx.scene.web.WebEngine
+import javafx.scene.web.WebView
+import mediathek.config.Konstanten
+import mediathek.gui.actions.DisposeDialogAction
+import mediathek.gui.actions.UrlHyperlinkAction
+import mediathek.gui.dialog.ButtonFlowPanel
+import mediathek.gui.dialog.ButtonPanel
+import mediathek.mainwindow.MediathekGui
+import mediathek.tool.EscapeKeyHandler
+import mediathek.tool.GuiFunktionen
+import mediathek.tool.Version
+import org.apache.logging.log4j.LogManager
+import java.awt.BorderLayout
+import java.awt.Frame
+import java.net.URI
+import java.net.URISyntaxException
+import java.util.*
+import javax.swing.JButton
+import javax.swing.JDialog
 
-import javax.swing.*;
-import java.awt.*;
-import java.util.Objects;
+class UpdateNotificationDialog(owner: Frame?, title: String?, private val version: Version) :
+    JDialog(owner, title, true) {
+    private val panel = UpdateNotificationPanel()
+    private var browser: WebView? = null
+    private lateinit var webEngine: WebEngine
 
-public class UpdateNotificationDialog extends JDialog {
-    private final Version version;
-    private final UpdateNotificationPanel panel = new UpdateNotificationPanel();
-    private WebView browser;
-    private WebEngine webEngine;
-
-    public UpdateNotificationDialog(Frame owner, String title, Version version) {
-        super(owner, title, true);
-        this.version = version;
-        setDefaultCloseOperation(JDialog.DISPOSE_ON_CLOSE);
-        EscapeKeyHandler.installHandler(this, this::dispose);
-
-        setupDialogInformation();
-        setupFxWebView();
-
-        var contentPane = getContentPane();
-        contentPane.setLayout(new BorderLayout());
-        contentPane.add(panel, BorderLayout.CENTER);
-        var buttonPanel = new ButtonPanel();
-        buttonPanel.add(createButtonPanel(), BorderLayout.EAST);
-        contentPane.add(buttonPanel, BorderLayout.SOUTH);
-        pack();
-        GuiFunktionen.centerOnScreen(this, false);
+    private fun createButtonPanel(): ButtonFlowPanel {
+        val pnl = ButtonFlowPanel()
+        val dlBtn = JButton("Download")
+        dlBtn.addActionListener {
+            try {
+                val uri = URI(Konstanten.ADRESSE_DOWNLOAD)
+                UrlHyperlinkAction.openURL(MediathekGui.ui(), uri.toString())
+                dispose()
+            } catch (e: URISyntaxException) {
+                logger.error(e)
+            }
+        }
+        val btn = JButton(DisposeDialogAction(this, "Schließen", "Dialog schließen"))
+        getRootPane().defaultButton = btn
+        pnl.add(dlBtn)
+        pnl.add(btn)
+        return pnl
     }
 
-    private ButtonFlowPanel createButtonPanel() {
-        ButtonFlowPanel pnl = new ButtonFlowPanel();
-        JButton btn = new JButton(new DisposeDialogAction(this, "Schließen", "Dialog schließen"));
-        getRootPane().setDefaultButton(btn);
-        pnl.add(btn);
-        return pnl;
+    private fun setupFxWebView() {
+        Platform.runLater {
+            browser = WebView()
+            val scene = Scene(browser)
+            webEngine = browser!!.engine
+            webEngine.load(Objects.requireNonNull(Konstanten.WEBSITE_BASE_URL.resolve("changelogs")).toString())
+            panel.fxPanel.scene = scene
+        }
     }
 
-    private void setupFxWebView() {
-        Platform.runLater(() -> {
-            browser = new WebView();
-            Scene scene = new Scene(browser);
-            webEngine = browser.getEngine();
-            webEngine.load(Objects.requireNonNull(Konstanten.WEBSITE_BASE_URL.resolve("changelogs")).toString());
-
-
-            panel.getFxPanel().setScene(scene);
-        });
+    private fun setupDialogInformation() {
+        val label = ("MediathekView " + version + " ist verfügbar - "
+                + "Sie haben Version " + Konstanten.MVVERSION)
+        panel.releaseInfoLabel.text = label
     }
 
-    private void setupDialogInformation() {
-        String label = "MediathekView " + version + " ist verfügbar - "
-                + "Sie haben Version " + Konstanten.MVVERSION;
-        panel.getReleaseInfoLabel().setText(label);
+    companion object {
+        private val logger = LogManager.getLogger()
+    }
 
+    init {
+        defaultCloseOperation = DISPOSE_ON_CLOSE
+        EscapeKeyHandler.installHandler(this) { dispose() }
+        setupDialogInformation()
+        setupFxWebView()
+        val contentPane = contentPane
+        contentPane.layout = BorderLayout()
+        contentPane.add(panel, BorderLayout.CENTER)
+        val buttonPanel = ButtonPanel()
+        buttonPanel.add(createButtonPanel(), BorderLayout.EAST)
+        contentPane.add(buttonPanel, BorderLayout.SOUTH)
+        pack()
+        GuiFunktionen.centerOnScreen(this, false)
     }
 }